+++ source/labyrinths/zooofchimeras.js
@@ -225,7 +225,7 @@ module.exports = new LabyrinthTemplate("Zoo of Chimeras",
 		// Labyrinth Particulars - more customized
 		"Event": ["Door 1 or Door 2?", "Twin Pedestals", "Imp Contract Faire", "Free Gold?", "Gear Collector", "The Score Beggar", "Apple Pie Wishing Well", "Workshop", "Merchant", "Rest Site", "Treasure"],
 		"Battle": ["Hawk Fight", "Wild Fire-Arrow Frogs", "Tortoise Fight"],
-		"Artifact Guardian": ["A Slimy Throneroom", "A windfall of treasure!"],
+		"Artifact Guardian": ["A Slimy Throneroom", "Brute Convention"],
 		"Final Battle": ["A Northern Laboratory", "Hall of Mirrors"],
 
 		// Labyrinth Infrastructure - less customized